Got it working
The issue was that you can't call provider.of inside a function that is passed to a button or some other
widget like that, you have to call it in the build function of the widget you want to register as a listener
of the provider. Disregard the commit message on my previous commit.