Working now with multiple pages
I got it so you can go to one page, come back and the text you entered on the pervious page is there. This is tricky because
the provider doesn't actually work across routes, so if you want to be able to access a provider across routes you have to create
the provider above any routes by making it a parent to the MaterialApp widget (it should be the widget returned by the build function of MyApp).